feat: correct REST semantics
